LayoutTests/fast/dom/HTMLElement/set-true.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="div1" contenteditable="false"></div>
<div id="div2" contenteditable="false">
    <p id="p2"></p>
</div>
<script>
description('When contentEditable is set to "true" string from "false" or inherit, element.contentEditable returns "true" and the element is editable.')

document.getElementById("div1").contentEditable = "true";
document.getElementById("p2").contentEditable = "true";

shouldBe('document.getElementById("div1").getAttribute("contentEditable")','"true"');
shouldBe('document.getElementById("div1").contentEditable', '"true"');
shouldBe('document.getElementById("div1").isContentEditable', 'true');
debug("FIXME: setContentEditable with true/false/inherit string is not working properly. https://bugs.webkit.org/show_bug.cgi?id=52058");
shouldBe('window.getComputedStyle(div1, "").getPropertyValue("-webkit-user-modify")', '"read-write"');

shouldBe('document.getElementById("p2").getAttribute("contentEditable")','"true"');
shouldBe('document.getElementById("p2").contentEditable', '"true"');
shouldBe('document.getElementById("p2").isContentEditable', 'true');
shouldBe('window.getComputedStyle(p2, "").getPropertyValue("-webkit-user-modify")', '"read-write"');

document.getElementById("div1").style.display= 'none';
document.getElementById("div2").style.display= 'none';
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
