third_party/blink/web_tests/fast/dom/HTMLElement/set-true-expected.txt

When contentEditable is set to "true" string from "false" or inherit, element.contentEditable returns "true" and the element is editable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("div1").getAttribute("contentEditable") is "true"
PASS document.getElementById("div1").contentEditable is "true"
PASS document.getElementById("div1").isContentEditable is true
FIXME: setContentEditable with true/false/inherit string is not working properly. https://bugs.webkit.org/show_bug.cgi?id=52058
PASS window.getComputedStyle(div1, "").getPropertyValue("-webkit-user-modify") is "read-write"
PASS document.getElementById("p2").getAttribute("contentEditable") is "true"
PASS document.getElementById("p2").contentEditable is "true"
PASS document.getElementById("p2").isContentEditable is true
PASS window.getComputedStyle(p2, "").getPropertyValue("-webkit-user-modify") is "read-write"
PASS successfullyParsed is true

TEST COMPLETE
